main func and Graph methods are separed to different files:p4c-graphs.cpp,graphs.cpp
