+++ backends/graphs/CMakeLists.txt
@@ -16,6 +16,7 @@ set (GRAPHS_SRCS
   graphs.cpp
   controls.cpp
   parser.cpp
+  p4c-graphs.cpp
   )
 
 set (GRAPHS_HDRS
